backend/requirements.txt
fastapi[all]
uvicorn[standard]==0.29.0
requests
supabase
python-jose[cryptography]==3.3.0
stripe==9.3.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
psycopg2-binary

# Database (Supabase uses PostgreSQL)
asyncpg==0.29.0
sqlalchemy==2.0.30

# Supabase Auth (if verifying JWTs manually)
python-jose[cryptography]==3.3.0

# Stripe payments
stripe==9.3.0

# Redis
redis==5.0.4

# HTTP client for Gemini API
httpx==0.27.0

# Input validation / schema
pydantic==2.7.1
pydantic-settings==0.2.2

# Utility
loguru==0.7.2

# Optional: testing and dev tools
pytest==8.2.1
httpx[http2]==0.27.0
pytest-asyncio==0.23.6
locust==2.27.0  # for load testing

gunicorn
gotrue

# New additions
watchfiles
supabase-auth==0.1.2
